add fixes

+++ database/migrations/2025_10_14_105026_add_per_restore_flag.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -15,13 +14,13 @@ return new class extends Migration
             if (Schema::hasColumn('skills', 'per_restore')) {
                 return;
             }
-            $table->tinyInteger('per_restore')->default(0)->after('vigor');
+            $table->unsignedTinyInteger('per_restore')->default(0)->after('vigor');
         });
-        Schema::table('feats', function ($table) {
+        Schema::table('feats', function (Blueprint $table) {
             if (Schema::hasColumn('feats', 'per_restore')) {
                 return;
             }
-            $table->tinyInteger('per_restore')->default(0)->after('per_day');
+            $table->unsignedTinyInteger('per_restore')->default(0)->after('per_day');
         });
     }
 
@@ -30,11 +29,15 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('skills', function (Blueprint $table) {
-            $table->dropColumn('per_restore');
-        });
-        Schema::table('feats', function (Blueprint $table) {
-            $table->dropColumn('per_restore');
-        });
+        if (Schema::hasColumn('skills', 'per_restore')) {
+            Schema::table('skills', function (Blueprint $table) {
+                $table->dropColumn('per_restore');
+            });
+        }
+        if (Schema::hasColumn('feats', 'per_restore')) {
+            Schema::table('feats', function (Blueprint $table) {
+                $table->dropColumn('per_restore');
+            });
+        }
     }
 };
